+++ src/components/Citation.vue
@@ -7,6 +7,7 @@ import type { PointOfContact as Contact, Identifier } from '../types/iso'
 
 import SectionTitle from './SectionTitle.vue'
 import Freetext from './Freetext.vue'
+import GuidanceText from './GuidanceText.vue'
 import Link from './Link.vue'
 
 import organisationsData from '../data/organisations.json'
@@ -158,11 +159,12 @@ watch(
         >
           Copy to input
         </button>
-        <em class="text-black dark:text-white"
-          >Click to copy this citation, with
-          <a
+        <GuidanceText
+          >Click to copy this citation (with
           <Link
             href="https://gitlab.data.bas.ac.uk/felnne/zap/-/blob/main/src/utils/crosscite.ts#L66"
+            >some modifications</Link
+          >) into the input below.</GuidanceText
         >
       </div>
     </div>