+++ src/components/GuidanceText.vue
@@ -1,0 +1,3 @@
+<template>
+  <p class="text-emerald-600"><slot></slot></p>
+</template>