+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Skipping `lint` CI task for tags in CI
 - Tweaking Prologue component language
 - Refactoring links into a `Link` base component
+- Refactoring guidance text into a `GuidanceText` base component
 
 ## [0.4.0] - 2023-12-27